backend/requirements.txt
fastapi
uvicorn
requests
python-multipart
pypdf
dotenv
pdfminer.high_level